Remove "geographicLocation" from mandatory metadata and add "datasetPointOfContact"
schema of datasetPointOfContact has been simplified, a contactPoint
could only have a role if it was a "individual", but ISO19115 seems to
allow a role for organisations too. So now possible.

The old "oneOf" in the schema also enforced users to give a "contactType",
which was oddly the only property mandatory. Now none are mandatory,
except "contactName" and "emailAddress". Geonovum also enforces in their
profile the "role", so perhaps that should be added, but this would break
files that are upgraded from v1.0 so I think we can live w/o this.