Address are flattened out + it's an array
fixes #51

Flattened out because many countries have very different ways to describe
an address, so better to leave it free. If one large string identifies a
building then it's fine.

The array is used when a Building, eg in LoD1.2, has several addresses.
This is common in NL for instance, and we don't always have access to
each BuildingUnit, so we just want to list all the addresses.