Harmonise all link to files with the property "url" (instead of "uri" and "url")
URL is better, and we want a HTTP URL so that Extension files can be
used by everyone.

The small schema to validate Extension files is updated, the examples,
and the specs.

Further, it shouldn't affect anything.

The small schema accepts "uri" or "url", and cjval will not complain if
either of these is present.