Change wording related to CityGML v3.0 conformance, based on OGC demands
So CityJSON v2.0 is no more conformant (there is no test suite available)
but implements a subset of v3.0 and has its own extension mechanism
(which is akin to ADEs).

Notice that the conformance page on the cityjson.org website will be
changed and since conformance cannot be claimed. It'll be "implementation"
and the URL will be https://cityjson.org/citygml/v30/ (to be changed soon)